Conflicts:
	Hystorical/Assets/Levels/UFO/Prefabs/Cow.prefab
	Hystorical/Assets/Levels/UFO/Prefabs/UFO.prefab
	Hystorical/Assets/Resources/NPC.prefab
	Hystorical/Assets/temp_Declan.unity

+++ Hystorical/Assets/Scripts/PersonController.cs
@@ -11,7 +11,7 @@ public class PersonController : MovementController {
 	
 	public float NPCFollowFactor = 1f;
 	public Vector3 NPCGoal = Vector3.zero;
-	public int NPCIntelligence = 0;
+	public float NPCIntelligence = 0;
     private int NPCChangeMind = 0;
     
     protected override void SetupTarget()
